refactor: modularize app utilities and components
